Add option to limit search by past months or past years
There is now a mutually exclusive group of arguments to limit the bibtex
publication age to the last `-m/--max_months` or `-y/--max_years`.

These entries are transformed into a `search_limit`, which is the oldest
year the entry can be published in. Note: max_months is rounded UP to
the search limit year.

Fixes #12